Add files via upload
A simple Flutter application that utilizes the video_player package to display a short video, which is played based on a random outcome. When the "Start" button is pressed, the Random function is used to determine whether the video to be played represents "Heads" or "Tails," and the appropriate video is then displayed. Additionally, the corresponding text result is shown at the top of the screen. The application also includes a static background image.